+++ src/Menus/Options/ApparatusOption/components/index.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 
 import { nudgeHarpStrataByApparatus } from '../nudgeHarpStrataByApparatus'
 import { partiallyApplyNudgeFunction } from '../../helpers'
-import { OptionContainer } from '../../OptionContainer'
+import { OptionContainer } from '../../../../components/option'
 
 export const ApparatusOption = (): React.ReactElement => {
   const [activeHarpStrata, setActiveHarpStrata] = useGlobal('activeHarpStrata')
