--- src/Menus/Options/OptionContainer/index.ts
@@ -1,1 +1,0 @@
-export { OptionContainer } from './components'
