--- src/Menus/Options/OptionContainer/types/index.ts
@@ -1,5 +1,0 @@
-export type OptionContainerProps = {
-  readonly title: string
-  readonly optionId: string
-  readonly nudgeFunction: (arg0: 'UP' | 'DOWN') => void
-}
